Shortened savegame loader
Changed the savegame loader so that it directly uses the filename in Gui
instead of searching for files in the savegame folder (which caused some
minor bugs). Since the filename string is specified in Gui, I also changed
the Autosaver to use that filename by adding it to the constructor. This
way, the location of the savegame can be easily changed in both the loader
and writer by just changing the filename in Gui.

< feature, organisation, bug