Additional improvements
Changed the GameTests to suit the new methods in Game. Added some
JavaDoc and made some other small code changes. checkMatches is
now executed after one or more matches are cleared so that matches
that are created by dropping down jewels or adding new jewels are
detected and cleared as well.

< feature, documentation