Added different Jewel types
Made Jewel abstract and added NormalJewel, HorizontalSuperJewel and
VerticalSuperJewel as concrete Jewels. Modified the JAXB-annotations
so that the game can still be correctly read and written. Added the
method getMatchCoordinates() to Jewel and its concrete implementations
so that the strategy pattern is used. This means that clearing supergems
should be working in theory. Support for creating supergems still needs
to be added.

< feature, organisation