Adjusted ImgLoader and Gui
I have changed the ImgLoader's hashmap to use ImageIcon instead of
BufferedImages since BufferedImages do not work work with animated gifs.
The GUI has been adjusted accordingly. The highlight jewel has been
destroyed and removed. After hours of trying to get it to work with gifs,
it does not seem to work. We are using buttons anyway so you can already
see which jewel is selected.