Supergem clearing + correct images
Both horizontal and vertical supergems are now cleared (the row/column
is removed) when a match is made with one of them. Also improved the
Colour handling in Jewel and in the Gui. The colour of a jewel can
only be set to a basic colour, so that jewels can still be correctly
compared to other jewels (a powergem has the same colour as a regular
gem). Added a method 'getImageColour()' to Jewel which returns the
correct enum of the gem for representation in the gui.

< feature