Fixed test + improved Board.setGrid()
Fixed the GameTest that was broken due to changes in the default
grid in Board. Also made the test cases use their own board instead
of the default one, to prevent these same type of errors in the
future. Lastly, improved Board.setGrid() so that the board dimensions
get updated when the grid is set.

< testing, bug, feature